move schedule updating mechanic out of TaxiScheduler
